% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PuntosCercanos.R
\name{PuntosCercanos}
\alias{PuntosCercanos}
\title{Calcula los puntos cercanos de dos vectores}
\usage{
PuntosCercanos(x, y)
}
\arguments{
\item{x}{primer vector (M dimensiones es soportado)}

\item{y}{segundo vector (M dimensiones es soportado)}
}
\value{
Índice devuelto
}
\description{
Dados dos vectores de n puntos , calcula los puntos más cercanos según distancia Euclídea y
 provee los índices y valores de ellos
}
\examples{
# Para calcular los puntos más cercanos de dos vectores de 5 puntos y 2 dimensiones sería:
x1 <- matrix(runif(50,0,2),nrow=5,ncol=2)
x2 <- matrix(runif(50,0,2),nrow=5,ncol=2)
PuntosCercanos(x1,x2)

#La función soporta M-dimensiones, por ejemplo para calcular
#la distancia en dos vectores de 5 puntos y 4 dimensiones sería:

x1 <- matrix(runif(20,0,2),nrow=5,ncol=4)
x2 <- matrix(runif(20,0,2),nrow=5,ncol=4)
PuntosCercanos(x1,x2)
}

